+++ lib/api/workday/dto/workday_for_employee_dto.dart
@@ -5,7 +5,7 @@ import 'package:give_job/api/work_time/dto/work_time_dto.dart';
 class WorkdayForEmployeeDto {
   final int id;
   final int number;
-  final int hours;
+  final double hours;
   final double money;
   final String plan;
   final String note;
@@ -27,7 +27,7 @@ class WorkdayForEmployeeDto {
     return WorkdayForEmployeeDto(
       id: json['id'] as int,
       number: json['number'] as int,
-      hours: json['hours'] as int,
+      hours: json['hours'] as double,
       money: json['money'] as double,
       plan: json['plan'] as String,
       note: json['note'] as String,
